# Decode Jpeg into Rgb24 when pixel type not specified