--- a/src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromCmykBasic.cs
+++ b/src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromCmykBasic.cs
@@ -18,36 +18,6 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder.ColorConverters
             public override void ConvertToRgbInplace(in ComponentValues values) =>
                 ConvertCoreInplace(values, this.MaximumValue);
 
-            internal static void ConvertCore(in ComponentValues values, Span<Vector4> result, float maxValue)
-            {
-                ReadOnlySpan<float> cVals = values.Component0;
-                ReadOnlySpan<float> mVals = values.Component1;
-                ReadOnlySpan<float> yVals = values.Component2;
-                ReadOnlySpan<float> kVals = values.Component3;
-
-                var v = new Vector4(0, 0, 0, 1F);
-
-                var maximum = 1 / maxValue;
-                var scale = new Vector4(maximum, maximum, maximum, 1F);
-
-                for (int i = 0; i < result.Length; i++)
-                {
-                    float c = cVals[i];
-                    float m = mVals[i];
-                    float y = yVals[i];
-                    float k = kVals[i] / maxValue;
-
-                    v.X = c * k;
-                    v.Y = m * k;
-                    v.Z = y * k;
-                    v.W = 1F;
-
-                    v *= scale;
-
-                    result[i] = v;
-                }
-            }
-
             internal static void ConvertCoreInplace(in ComponentValues values, float maxValue)
             {
                 Span<float> c0 = values.Component0;