… src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromRgbVector8.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromYCbCrAvx2.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromYCbCrBasic.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromYCbCrVector4.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromYCbCrVector8.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromYccKAvx2.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromYccKBasic.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromYccKVector8.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.VectorizedJpegColorConverter.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/HuffmanScanBuffer.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/SpectralConverter{TPixel}.cs
… src/ImageSharp/Formats/Jpeg/JpegDecoder.cs
… src/ImageSharp/Formats/Jpeg/JpegDecoderCore.cs
… src/ImageSharp/Formats/Tiff/Compression/Decompressors/JpegTiffCompression.cs
… src/ImageSharp/Memory/Buffer2D{T}.cs
… src/ImageSharp/Memory/DiscontiguousBuffers/MemoryGroupExtensions.cs
… src/ImageSharp/PixelFormats/PixelImplementations/PixelOperations/Rgb24.PixelOperations.cs
… src/ImageSharp/PixelFormats/PixelImplementations/PixelOperations/Rgba32.PixelOperations.cs
… src/ImageSharp/PixelFormats/PixelOperations{TPixel}.cs
… src/ImageSharp/PixelFormats/Utils/Vector4Converters.cs
… src/ImageSharp/Processing/Processors/Transforms/Resize/ResizeProcessor{TPixel}.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/CmykColorConversion.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/ColorConversionBenchmark.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/DecodeJpeg_ImageSpecific.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/GrayscaleColorConversion.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/RgbColorConversion.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/YCbCrColorConversion.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/YccKColorConverter.cs
… tests/ImageSharp.Benchmarks/Config.cs