### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Finish the plan defined in #1410:
- Remove packing from JpegColorConverters and utilize #1462 `PixelOperations.PackFromRgbPlanes` instead
- Return `Image<Rgb24>` from non-generic Decode
- `ResizeProcessor<T>`:  omit alpha premultiplication if the pixel type has no alpha channel

Fixes #1410, fixes #1121.

### Performance

Did usual run of `LoadResizeSaveStressBenchmarks` on 10 core desktop. The overall effect on speed is within noise range, but there might be some 1%-ish improvement. I assume the reason is that #1462's packing actually doesn't make big difference to the packing code introduced in #1411.

But memory usage dropped significantly, which is the main goal here.

#### Before
```
|        Method | maxDegreeOfParallelism |       Mean |       Error |   StdDev | Ratio | RatioSD |     Gen 0 |     Gen 1 |     Gen 2 |    Allocated |
|-------------- |----------------------- |-----------:|------------:|---------:|------:|--------:|----------:|----------:|----------:|-------------:|
| SystemDrawing |                      1 | 8,711.9 ms |   862.64 ms | 47.28 ms |  1.00 |    0.00 |         - |         - |         - |        25 KB |
|    ImageSharp |                      1 | 5,218.4 ms |   135.62 ms |  7.43 ms |  0.60 |    0.00 | 3000.0000 | 3000.0000 | 3000.0000 | 1,226,257 KB |
|               |                        |            |             |          |       |         |           |           |           |              |
| SystemDrawing |                      5 | 3,128.0 ms | 1,188.61 ms | 65.15 ms |  1.00 |    0.00 |         - |         - |         - |        26 KB |
|    ImageSharp |                      5 | 1,516.9 ms |   782.02 ms | 42.86 ms |  0.49 |    0.02 | 2000.0000 | 2000.0000 | 2000.0000 | 1,226,402 KB |
|               |                        |            |             |          |       |         |           |           |           |              |
| SystemDrawing |                     10 | 2,828.7 ms |   786.61 ms | 43.12 ms |  1.00 |    0.00 |         - |         - |         - |        29 KB |
|    ImageSharp |                     10 |   823.4 ms |    80.49 ms |  4.41 ms |  0.29 |    0.01 | 1000.0000 | 1000.0000 | 1000.0000 | 1,229,766 KB |
|               |                        |            |             |          |       |         |           |           |           |              |
| SystemDrawing |                     20 | 2,793.0 ms |   735.65 ms | 40.32 ms |  1.00 |    0.00 |         - |         - |         - |        30 KB |
|    ImageSharp |                     20 |   686.0 ms |   570.78 ms | 31.29 ms |  0.25 |    0.01 | 1000.0000 | 1000.0000 | 1000.0000 | 1,237,201 KB |
```

#### After
```
|        Method | maxDegreeOfParallelism |       Mean |      Error |   StdDev | Ratio | RatioSD |     Gen 0 |     Gen 1 |     Gen 2 |  Allocated |
|-------------- |----------------------- |-----------:|-----------:|---------:|------:|--------:|----------:|----------:|----------:|-----------:|
| SystemDrawing |                      1 | 8,757.5 ms | 1,109.5 ms | 60.81 ms |  1.00 |    0.00 |         - |         - |         - |      25 KB |
|    ImageSharp |                      1 | 4,868.9 ms |   258.8 ms | 14.18 ms |  0.56 |    0.00 | 2000.0000 | 2000.0000 | 2000.0000 | 776,094 KB |
|               |                        |            |            |          |       |         |           |           |           |            |
| SystemDrawing |                      5 | 3,222.5 ms |   913.8 ms | 50.09 ms |  1.00 |    0.00 |         - |         - |         - |      27 KB |
|    ImageSharp |                      5 | 1,397.8 ms |   624.2 ms | 34.21 ms |  0.43 |    0.02 |         - |         - |         - | 776,093 KB |
|               |                        |            |            |          |       |         |           |           |           |            |
| SystemDrawing |                     10 | 2,849.4 ms | 1,590.1 ms | 87.16 ms |  1.00 |    0.00 |         - |         - |         - |      29 KB |
|    ImageSharp |                     10 |   852.6 ms |   576.0 ms | 31.57 ms |  0.30 |    0.02 | 1000.0000 | 1000.0000 | 1000.0000 | 843,968 KB |
|               |                        |            |            |          |       |         |           |           |           |            |
| SystemDrawing |                     20 | 2,764.5 ms | 1,152.1 ms | 63.15 ms |  1.00 |    0.00 |         - |         - |         - |      30 KB |
|    ImageSharp |                     20 |   622.7 ms |   475.5 ms | 26.07 ms |  0.23 |    0.01 | 1000.0000 | 1000.0000 | 1000.0000 | 949,029 KB |
```

/cc @br3aker @JimBobSquarePants @saucecontrol